Changes to be committed: 	new file:   docker-compose.yml 	modified:   main.py 	deleted:    scan_results.json 	modified:   src/download_files.py 	modified:   src/s3_upload.py 	modified:   src/yara_scan.py

+++ src/download_files.py
@@ -1,4 +1,5 @@
 # download_files.py
+
 import requests
 from pathlib import Path
 from datetime import datetime
@@ -10,10 +11,17 @@ def download_file(date: datetime) -> Optional[Path]:
     filename = f"Virussign.{date.strftime('%Y.%m.%d')}.7z"  # Формат файла: Virussign.YYYY.MM.DD.7z
     url = f"{base_url}{folder}/{filename}"
 
+    zip_path = Path("downloads") / filename  # Сохраняем файл в папку 'downloads'
+    zip_path.parent.mkdir(exist_ok=True)
+
+    # Проверка, если файл уже существует
+    if zip_path.exists():
+        print(f"Файл {filename} уже загружен ранее.")
+        return zip_path
+
     response = requests.get(url, stream=True)
 
     if response.status_code == 200:
-        zip_path = Path(filename)
         with open(zip_path, "wb") as file:
             for chunk in response.iter_content(chunk_size=8192):
                 file.write(chunk)
@@ -24,11 +32,11 @@ def download_file(date: datetime) -> Optional[Path]:
     return None
 
 def extract_files(zip_path: Path) -> Path:
-    extract_dir = Path("extracted_files")
-    extract_dir.mkdir(exist_ok=True)
+    extract_dir = Path("extracted_files") / zip_path.stem  # Директория для распаковки каждого архива
+    extract_dir.mkdir(parents=True, exist_ok=True)
 
     import py7zr  # Библиотека для работы с 7z архивами
     with py7zr.SevenZipFile(zip_path, mode='r', password='infected') as z:
         z.extractall(path=extract_dir)
-    print(f"Файлы извлечены в директорию {extract_dir}")
+    print(f"Файлы из {zip_path.name} извлечены в директорию {extract_dir}")
     return extract_dir
